Split eth and erc20 escrow tests into separate files
